pyproject.toml
[tool.poetry]
name = "mapanare"
version = "0.1.0"
description = "Mapanare a GIS aplication"
authors = ["Roger Zavarce <rogerzavarce@sipay.es>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
django-leaflet = "^0.28.0"
django-geojson = "^3.2.0"
django-timezone-field = "^4.1.2"
django-datetime-widget = "^0.9.3"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.1.0"}
jsonfield = "^3.1.0"
Pillow = "^8.2.0"
django-baton = "^2.1.5"
django-base-url = "^2020.12.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
